Move transform functions from utils.py to transforms.py
