+++ custom_components/ef_ble/eflib/devices/delta3.py
@@ -1,7 +1,8 @@
 from ..entity import controls
 from ..pb import pd335_sys_pb2
 from ..props import pb_field
-from . import _delta3_base, delta3_classic
+from ..props.transforms import flow_is_on
+from . import delta3_classic
 
 pb = delta3_classic.pb
 
@@ -12,7 +13,7 @@ class Device(delta3_classic.Device):
     SN_PREFIX = (b"P231",)
     NAME_PREFIX = "EF-D3"
 
-    usb_ports = pb_field(pb.flow_info_qcusb1, _delta3_base.flow_is_on)
+    usb_ports = pb_field(pb.flow_info_qcusb1, flow_is_on)
 
     @controls.switch(usb_ports)
     async def enable_usb_ports(self, enabled: bool):
