+++ custom_components/ef_ble/eflib/devices/delta3_max_plus.py
@@ -3,14 +3,10 @@ from collections.abc import Sequence
 from ..entity import controls
 from ..pb import pd335_sys_pb2
 from ..props import pb_field, repeated_pb_field_type
+from ..props.transforms import flow_is_on, out_power
 from . import delta3_plus
-from ._delta3_base import flow_is_on, out_power
 
 pb = delta3_plus.pb
-
-
-def _out_power(x) -> float:
-    return -round(x, 2) if x != 0 else 0
 
 
 class _ACPortPower(repeated_pb_field_type(pb.pow_get_ac_out_list.pow_get_ac_out_item)):
